server_python/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.1
asyncpg>=0.29.0
firebase-admin>=6.5.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
httpx>=0.27.0  # For testing or external requests
pytest>=8.0.0
